@@ -1,16 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿namespace SandBox.ExtensionMethods;
 
-namespace SandBox.ExtensionMethods
+public static class StringExtensions
 {
-    public static class StringExtensions
+    public static void DumpThread(this string value)
     {
-        public static void DumpThread(this string value)
-        {
-            Console.WriteLine($"[{DateTime.Now:hh:mm:ss.fff}] {value}: TID:{Thread.CurrentThread.ManagedThreadId}");
-        }
+        Console.WriteLine($"[{DateTime.Now:hh:mm:ss.fff}] {value}: TID:{Thread.CurrentThread.ManagedThreadId}");
     }
 }
